+++ R/weathercan-package.R
@@ -75,6 +75,7 @@ NULL
   options(
     weathercan.urls.weather = "https://climate.weather.gc.ca/climate_data/bulk_data_e.html",
     weathercan.urls.normals = "https://climate.weather.gc.ca/climate_normals/bulk_data_e.html",
+    weathercan.urls.normals_1991_2020 = "https://climate.weather.gc.ca/climate_normals/download/1991_e.html?area=canada_wide&product%5B%5D=data&product%5B%5D=station_inventory",
     # Download from google drive: https://stackoverflow.com/a/50533232/3362144
     weathercan.urls.stations = paste0(
       "https://collaboration.cmc.ec.gc.ca/cmc/climate/",