+++ tests/testthat/test_08_normals.R
@@ -1,24 +1,5 @@
 # normals_html() ------------------------------------------------------------
 
-test_that("normals_html() correctly retrieves request 1991-2020", {
-  skip("New normals not ready yet")
-  skip_on_cran()
-  skip_if_offline()
-
-  expect_silent(
-    nd <- normals_html(
-      station_id = 3471,
-      climate_id = "5010480",
-      normals_years = "1991-2020",
-      prov = "MB"
-    )
-  )
-
-  expect_s3_class(nd, "response")
-  expect_false(httr::http_error(nd))
-  expect_gt(length(nd$content), 10000)
-})
-
 test_that("normals_html() correctly retrieves request 1981-2010", {
   skip_on_cran()
   skip_if_offline()
@@ -260,66 +241,33 @@ test_that("normals_format()/frost_format() format data to correct class", {
 
 
 # normals_dl() ------------------------------------------------------------
-test_that("normals_dl() downloads normals/frost dates as tibble - single", {
+test_that("normals_dl() - single - 1981", {
   skip_on_cran()
   skip_if_offline()
 
   memoise::forget(normals_html) # Reset cache so we can test fully
 
-  # 1981-2010
   expect_silent(
-    nd1 <- normals_dl(climate_id = "5010480", normals_years = "1981-2010")
-  ) %>%
-    expect_s3_class("tbl_df")
-
-  # 1971-2000
-  expect_silent(
-    nd2 <- normals_dl(climate_id = "5010480", normals_years = "1971-2000")
-  ) %>%
-    expect_s3_class("tbl_df")
-
-  expect_snapshot_value(nd1, style = "json2", tolerance = 0.001)
-  expect_snapshot_value(nd2, style = "json2", tolerance = 0.001)
-
-  skip("New normals not ready yet")
-  # 1991-2020
-  expect_silent(
-    nd1 <- normals_dl(climate_id = "5010480", normals_years = "1991-2020")
-  ) %>%
-    expect_s3_class("tbl_df")
+    nd <- normals_dl(climate_id = "5010480", normals_years = "1981-2010")
+  )
+  expect_s3_class(nd, "tbl_df")
+  expect_snapshot_value(nd, style = "json2", tolerance = 0.001)
 })
 
-test_that("normals_dl() downloads normals/frost dates as tibble - multi 1991", {
-  skip("New normals not ready yet")
+test_that("normals_dl() - single - 1971", {
   skip_on_cran()
   skip_if_offline()
 
-  expect_silent(
-    nd <- normals_dl(
-      climate_id = c("2403500", "5010480", "1096450"),
-      normals_years = "1991-2020"
-    )
-  ) %>%
-    expect_s3_class("tbl_df")
+  memoise::forget(normals_html) # Reset cache so we can test fully
 
-  expect_equal(nrow(nd), 3)
-  expect_s3_class(tidyr::unnest(nd, normals), "data.frame")
-  expect_s3_class(tidyr::unnest(nd, frost), "data.frame")
-  expect_length(
-    tidyr::unnest(nd, normals) %>%
-      dplyr::pull(climate_id) %>%
-      unique(),
-    3
+  expect_silent(
+    nd <- normals_dl(climate_id = "5010480", normals_years = "1971-2000")
   )
-  expect_length(
-    tidyr::unnest(nd, frost) %>%
-      dplyr::pull(climate_id) %>%
-      unique(),
-    0
-  )
+  expect_s3_class(nd, "tbl_df")
+  expect_snapshot_value(nd, style = "json2", tolerance = 0.001)
 })
 
-test_that("normals_dl() downloads normals/frost dates as tibble - multi 1981", {
+test_that("normals_dl() - multi 1981", {
   skip_on_cran()
   skip_if_offline()
 
@@ -348,7 +296,7 @@ test_that("normals_dl() downloads normals/frost dates as tibble - multi 1981", {
   )
 })
 
-test_that("normals_dl() downloads normals/frost dates as tibble - multi 1971", {
+test_that("normals_dl() - multi 1971", {
   skip_on_cran()
   skip_if_offline()
 
@@ -409,23 +357,21 @@ test_that("normals_dl() gets extreme wind chill correctly", {
   skip_on_cran()
   skip_if_offline()
 
-  expect_message(nd <- normals_dl(climate_id = "2100517"), "current normals")
+  expect_silent(
+    nd <- normals_dl(climate_id = "2100517", normals_years = "1981-2010")
+  )
   expect_s3_class(nd, "tbl_df")
 })
 
 test_that("normals_dl() multiple weird stations", {
   skip_on_cran()
   skip_if_offline()
-  expect_message(
-    nd <- normals_dl(climate_ids = c("301C3D4", "301FFNJ", "301N49A")),
-    "current normals"
+  expect_silent(
+    nd <- normals_dl(
+      climate_ids = c("301C3D4", "301FFNJ", "301N49A"),
+      normals_years = "1981-2010"
+    )
   )
 
   expect_snapshot_value(nd, style = "json2", tolerance = 0.001)
-})
-test_that("normals_dl validates normals_years", {
-  expect_error(
-    normals_dl("5010480", normals_years = "abcd"),
-    "'normals_years' must be either 'current' or a text string in the format YYYY-YYYY e.g., '1981-2010'"
-  )
 })